#72 申请添加好友调通

+++ seen-micro-database/seen-chat/src/main/java/io/github/seenings/apply/po/UserApplyTradePO.java
@@ -32,7 +32,7 @@ public class UserApplyTradePO {
     /**
      * 交易ID
      */
-    private Integer tradeId;
+    private Long tradeId;
 
     /**
      * 创建时间
